update navbar styling
